chore(fe): rename hook files
